Several tests for OrderService implemented. Code changes according to PR comments. Other minor test improvements. Product and ProductOrder Quantity property type modified to int and Price property modified to decimal according to previous PR

+++ EcommerceStore.Application/Models/ViewModels/ProductViewModel.cs
@@ -4,8 +4,8 @@ namespace EcommerceStore.Application.Models.ViewModels
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
-        public decimal Quantity { get; set; }
+        public decimal Price { get; set; }
+        public int Quantity { get; set; }
         public string Description { get; set; }
         public string Image { get; set; }
         public string BrandName { get; set; }
